+++ dwds/lib/src/servers/extension_backend.dart
@@ -8,12 +8,9 @@ import 'dart:io';
 import 'package:async/async.dart';
 import 'package:http_multi_server/http_multi_server.dart';
 import 'package:shelf/shelf_io.dart';
-import 'package:sse/server/sse_handler.dart';
 
+import '../handlers/socket_connections.dart';
 import 'extension_debugger.dart';
-
-final _sseHandler =
-    SseHandler(Uri.parse('/\$debug'), keepAlive: const Duration(seconds: 30));
 
 /// A backend for the Dart Debug Extension.
 ///
@@ -29,18 +26,22 @@ class ExtensionBackend {
   /// All subsequent calls to [close] will return this future.
   Future<void> _closed;
 
-  ExtensionBackend._(this.hostname, this.port, this._server);
+  ExtensionBackend._(
+      SocketHandler _socketHandler, this.hostname, this.port, this._server)
+      : connections = _socketHandler.connections;
 
   // Starts the backend on an open port.
-  static Future<ExtensionBackend> start(String hostname) async {
+  static Future<ExtensionBackend> start(
+      SocketHandler _socketHandler, String hostname) async {
     var server = await HttpMultiServer.bind(hostname, 0);
-    serveRequests(server, _sseHandler.handler);
-    return ExtensionBackend._(server.address.host, server.port, server);
+    serveRequests(server, _socketHandler.handler);
+    return ExtensionBackend._(
+        _socketHandler, server.address.host, server.port, server);
   }
 
   Future<void> close() => _closed ??= _server.close();
 
-  StreamQueue<SseConnection> connections = _sseHandler.connections;
+  final StreamQueue<SocketConnection> connections;
 
   Future<ExtensionDebugger> get extensionDebugger async =>
       ExtensionDebugger(await connections.next);