+++ dwds/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 5.1.0-dev
+
+- Support using WebSockets for the debug backend by passing
+  `useSseForDebugBackend: false` to `Dwds.start()`
+- Ensure we run main on a hot restart request even if no modules were
+  updated. 
+
 ## 5.0.0
 
 - Have unimplemented VM service protocol methods return the RPC error